Moved out reactivity from settings display
