+++ components/GardenPlanner.vue
@@ -7,8 +7,11 @@ import { useTakingScreenshot } from '~/stores/useIsTakingScreenshot'
 import useGarden from '~/stores/useGarden'
 import { storeToRefs } from 'pinia'
 import useHarvester from '~/stores/useHarvester'
-import useProcessor from '~/stores/useProcessor'
 import type { TUniqueTiles } from '~/assets/scripts/garden-planner/utils/garden-helpers'
+import { useSettingsCode } from '~/stores/useSettingsCode'
+import {  ItemType } from '~/assets/scripts/garden-planner/utils/garden-helpers'
+import {  getCropFromType } from '~/assets/scripts/garden-planner/imports'
+
 
 const display = ref<HTMLElement | null>(null)
 const isTakingScreenshot = useTakingScreenshot()
@@ -17,13 +20,88 @@ const isTakingScreenshot = useTakingScreenshot()
 const gardenHandler = useGarden()
 const harvester = useHarvester()
 const processor = useProcessor()
-const { isFullUpdateRequested } = storeToRefs(gardenHandler)
-
 
 watchEffect(() => {
   gardenHandler.update()
   harvester.harvester.simulateYield(gardenHandler.garden.uniqueTiles as TUniqueTiles, harvester.settings)
   processor.simulateProcessing(harvester.totalHarvest)
+})
+
+const garden = useGarden()
+const settingsCode = useSettingsCode()
+
+
+const starBaseChance = ref(0.25 + (harvester.settings.useStarSeeds ? 0.25 : 0) + (harvester.settings.level * 0.02))
+
+watchEffect(() => {
+  if (harvester.settings.level < 0)
+    harvester.settings.level = 0
+
+  starBaseChance.value = 0.25 + (harvester.settings.useStarSeeds ? 0.25 : 0) + (harvester.settings.level * 0.02)
+
+  starBaseChance.value = Math.min(1, starBaseChance.value)
+
+  if (harvester.settings.days === 'L')
+    harvester.settings.days = -1
+  else if (harvester.settings.days === 'M')
+    harvester.settings.days = 0
+  else if (harvester.settings.days < -1)
+    harvester.settings.days = -1
+
+  harvester.updateSettings({ ...harvester.settings })
+})
+
+
+function saveGarden() {
+  const saveString = garden.garden.saveSettings(harvester.settings, processor.settings)
+  settingsCode.set(saveString)
+}
+
+function loadGarden(saveString: string) {
+  const { harvesterOptions, processorSettings: loadedProcessorSettings } = garden.garden.loadSettings(saveString)
+
+  // updateHarvesterSettings(Object.assign({}, harvesterOptions))
+  harvester.updateSettings(Object.assign({}, harvesterOptions))
+  processor.updateSettings(Object.assign({}, loadedProcessorSettings))
+  processor.simulateProcessing(harvester.totalHarvest)
+}
+
+
+watchEffect(() => {
+  // set all isActive to false
+  for (const setting of processor.settings.cropSettings.values())
+    setting.isActive = false
+
+  for (const [cropId, data] of harvester.totalHarvest.crops) {
+    const cropSetting = processor.settings.cropSettings.get(cropId) ?? {
+      count: data.totalWithDeductions,
+      cropType: data.cropType,
+      isStar: data.isStar,
+      processAs: ItemType.Crop,
+      crafters: 1,
+      targetTime: 0,
+      isActive: true,
+      hasPreserve: (getCropFromType(data.cropType)?.conversionInfo.preserveProcessMinutes || 0) > 0,
+    }
+
+    cropSetting.count = data.totalWithDeductions
+
+    processor.settings.cropSettings.set(cropId, cropSetting)
+
+    processor.settings.cropSettings.get(cropId)!.isActive = true
+  }
+
+  saveGarden()
+})
+
+
+const { updateIsRequested } = storeToRefs(settingsCode)
+
+watch(updateIsRequested, () => {
+  if (updateIsRequested.value) {
+    loadGarden(settingsCode.code)
+    settingsCode.resetUpdateRequest()
+  }
 })
